REF: Drop the redundant float64 copy in the ordinal generator

+++ skordinal/datasets/_samples_generator.py
@@ -146,4 +146,4 @@ def make_ordinal_classification(
     thresholds = np.percentile(z, cut_percentiles)
     y = np.searchsorted(thresholds, z).astype(np.intp)
 
-    return X.astype(np.float64), y
+    return X, y
